rename module to connectors
