+++ singer_sdk/connectors/__init__.py
@@ -1,5 +1,5 @@
 """Module for SQL-related operations."""
 
-from .connector import SQLConnector
+from .sql import SQLConnector
 
 __all__ = ["SQLConnector"]
